requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.3
billboard.py==6.2.1
boto3==1.16.15
botocore==1.19.15
certifi==2020.11.8
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.1
django-environ==0.4.5
django-heroku==0.3.1
elasticsearch==7.9.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6

pylint==2.6.0


python-dateutil==2.8.1
python-dotenv==0.15.0
pyttsx3==2.90
pytz==2020.1
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
toml==0.10.2
urllib3==1.25.11
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
